# specs/004-subscription-tracking/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Subscription Tracking API
  description: API endpoints for managing subscription inventory with configurable categories
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
paths:
...
  /categories:
    get:
      summary: List all categories with subcategories
      description: Returns all categories with their nested subcategories. Used to populate dropdowns.
      operationId: listCategories
      parameters:
        - name: include_inactive
          in: query
          description: Include inactive (soft-deleted) categories
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of categories with subcategories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryWithSubcategories'

    post:
      summary: Create a new category
      description: Creates a new subscription category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Category name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /categories/{category_id}:
    parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer

    get:
      summary: Get category by ID
      operationId: getCategory
      responses:
        '200':
          description: Category details with subcategories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryWithSubcategories'
        '404':
          description: Category not found

    put:
      summary: Update category
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Category not found

    delete:
      summary: Delete category (soft delete)
      description: Soft deletes category. Fails if subscriptions use this category.
      operationId: deleteCategory
      responses:
        '204':
          description: Category deleted
        '400':
          description: Cannot delete - category in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
  /categories/{category_id}/usage:
    parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer

    get:
      summary: Get category usage count
      description: Returns count of subscriptions using this category
      operationId: getCategoryUsage
      responses:
        '200':
          description: Usage count
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription_count:
                    type: integer

  # ============================================
  # SUBCATEGORY ENDPOINTS
  # ============================================
  /categories/{category_id}/subcategories:
    parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer

    get:
      summary: List subcategories for a category
      operationId: listSubcategories
      parameters:
        - name: include_inactive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of subcategories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubcategoryResponse'

    post:
      summary: Create subcategory
      operationId: createSubcategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubcategoryCreate'
      responses:
        '201':
          description: Subcategory created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubcategoryResponse'
        '400':
          description: Subcategory name already exists in this category
  /subcategories/{subcategory_id}:
    parameters:
      - name: subcategory_id
        in: path
        required: true
        schema:
          type: integer

    put:
      summary: Update subcategory
      operationId: updateSubcategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubcategoryUpdate'
      responses:
        '200':
          description: Subcategory updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubcategoryResponse'

    delete:
      summary: Delete subcategory (soft delete)
      description: Soft deletes subcategory. Fails if subscriptions use it.
      operationId: deleteSubcategory
      responses:
        '204':
          description: Subcategory deleted
        '400':
          description: Cannot delete - subcategory in use
  /subcategories/{subcategory_id}/usage:
    parameters:
      - name: subcategory_id
        in: path
        required: true
        schema:
          type: integer

    get:
      summary: Get subcategory usage count
      operationId: getSubcategoryUsage
      responses:
        '200':
          description: Usage count
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription_count:
                    type: integer

  # ============================================
  # SUBSCRIPTION ENDPOINTS
  # ============================================
...
components:
  schemas:
...
    CategoryCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        display_order:
          type: integer
          default: 0
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        display_order:
          type: integer
        is_active:
          type: boolean
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        display_order:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CategoryWithSubcategories:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        display_order:
          type: integer
        is_active:
          type: boolean
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/SubcategoryResponse'

    # ============================================
    # SUBCATEGORY SCHEMAS
    # ============================================
    SubcategoryCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        display_order:
          type: integer
          default: 0
    SubcategoryUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        display_order:
          type: integer
        is_active:
          type: boolean
    SubcategoryResponse:
      type: object
      properties:
        id:
          type: integer
        category_id:
          type: integer
        name:
          type: string
        display_order:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ============================================
    # SUBSCRIPTION SCHEMAS
    # ============================================
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string